Propagate path via prop drilling rather than using asPath
