+++ components/Shell/Feedback/FeedbackForm.tsx
@@ -2,7 +2,6 @@
 
 import clsx from 'clsx';
 import { useMode, usePlatform } from 'context/state';
-import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { Mode, Platform } from 'types';
 
@@ -25,6 +24,7 @@ type FeedbackOption = {
 type Props = {
   title: string;
   options: FeedbackOption[];
+  path: string;
 };
 
 export const negativeFeedbackOptions: FeedbackOption[] = [
@@ -100,8 +100,7 @@ const sendFeedback = async (feedback: FeedbackData) => {
   }
 };
 
-export const FeedbackForm = ({ title, options }: Props) => {
-  const { asPath, isReady } = useRouter();
+export const FeedbackForm = ({ title, options, path }: Props) => {
   const { mode } = useMode();
   const { platform } = usePlatform();
   const pageTitle = window.document.title;
@@ -111,19 +110,12 @@ export const FeedbackForm = ({ title, options }: Props) => {
   const [comment, setComment] = useState<string>();
   const [opacity, setOpacity] = useState('opacity-0');
   const [feedbackSubmitted, setFeedbackSubmitted] = useState(false);
-  const [pathname, setPathname] = useState<string>('');
 
   useEffect(() => {
     setTimeout(() => {
       setOpacity('opacity-100');
     }, 100);
   }, []);
-
-  useEffect(() => {
-    if (isReady) {
-      setPathname(asPath);
-    }
-  }, [isReady, asPath]);
 
   return feedbackSubmitted ? (
     <h3 className="mb-10 mt-5">Thanks for the feedback!</h3>
@@ -247,7 +239,7 @@ export const FeedbackForm = ({ title, options }: Props) => {
               email,
               mode,
               option: selectedOption.title,
-              path: pathname,
+              path,
               platform,
               title: removeTrailingDocs(pageTitle)
             });
