+++ components/Shell/index.ts
@@ -2,7 +2,7 @@
 
 export * from 'components/Shell/Feedback';
 export * from 'components/Shell/TopNavigation/TopNav';
-export * from 'components/Shell/SideNav';
+export * from 'components/Shell/SideNavigation/SideNav';
 export * from 'components/Shell/Aside';
 export * from 'components/Shell/Footer';
 export * from 'components/Shell/PageHistory';
