+++ pages/[...slug]/index.tsx
@@ -112,10 +112,12 @@ export const getStaticProps: GetStaticProps<object, Params> = async ({
 
 export default function Page({
   source,
-  frontmatter
+  frontmatter,
+  path
 }: {
   source: string;
   frontmatter: any;
+  path: string;
 }) {
   const router = useRouter();
 
@@ -131,7 +133,8 @@ export default function Page({
   const updatedConfig: Config = {
     ...config,
     variables: {
-      frontmatter
+      frontmatter,
+      path
     }
   };
 
@@ -139,7 +142,7 @@ export default function Page({
   return (
     <div className="mt-[6.5rem] flex grow flex-row justify-center lg:mt-[3.75rem]">
       <div className="flex max-w-[100rem] grow flex-row justify-center">
-        <SideNav />
+        <SideNav path={path} />
         <div className="grow">
           <div id="skip-nav" />
           {Markdoc.renderers.react(content, React, { components })}
